fix: double negative sign within the std::ostream& polynomial operator and display function within the cpp files

+++ Advanced Calculator/Functions/Polynomial.hpp
@@ -56,7 +56,7 @@ public:
     Polynomial integrate() const;
     
     // Algebraic Operations
-    double evaluate(double);
+    float evaluate(float);
     
 private:
     std::vector<IndeterminantC::Indeterminant> expressions;
